Skip unsupported directory/file names in list
GCS supports objects of path-type
e.g. <bucket>/<path1>//<path2>
and treats it different from
<bucket>/<path1>/<path2>.
GCSFuse being a POSIX-compliant file-system
can only support the latter and
throws error on former. This error can
disallow the listing of all directories in the
parent directory i.e. <path1>.
The current change ignores the listing of
prefixes (directory names) which are empty such
as "/" above to ignore the error and logs the
above event as a warning.

Similarly, GCS supports <bucket>/<path1>/./<path2>
and <bucket>/<path1>/../<path2> but these have
special reserved meaning in linux filesystem.
So, ignoring these two prefixes as well.

Similarly, linux filesystem does not support
'\0' in names of files or directories while it
is supported in GCS object names. So, ignoring
those as well in the GCS list call output.